refactor(repo): re use vault table from tangle-dapp to tangle-cloud
