+++ apps/tangle-cloud/src/app/providers.tsx
@@ -4,7 +4,6 @@ import {
   AppEvent,
   WebbProvider,
 } from '@tangle-network/api-provider-environment';
-import { RestakeContextProvider } from '@tangle-network/tangle-shared-ui/context/RestakeContext';
 import { UIProvider } from '@tangle-network/ui-components';
 import { type PropsWithChildren, type ReactNode } from 'react';
 import type { State } from 'wagmi';
@@ -28,7 +27,7 @@ const Providers = ({
         wagmiInitialState={wagmiInitialState}
       >
         <PolkadotApiProvider>
-          <RestakeContextProvider>{children}</RestakeContextProvider>
+          {children}
         </PolkadotApiProvider>
       </WebbProvider>
     </UIProvider>
