+++ libs/tangle-shared-ui/src/utils/createVaultMap.ts
